+++ examples/properties.cpp
@@ -37,17 +37,5 @@ for (auto e : mesh.edges())
 mesh.remove_edge_property(edge_points);
 //! [edge-properties]
 
-//! [global-properties]
-// get or add object property
-auto markers = mesh.object_property<std::vector<int>>("o:regions");
-markers[0].push_back(0);
-markers[0].push_back(1);
-
-// ...
-// do something useful
-// ...
-
-mesh.remove_object_property(markers);
-//! [global-properties]
 }
 // clang-format on