+++ src/pmp/SurfaceMesh.cpp
@@ -10,8 +10,6 @@ namespace pmp {
 
 SurfaceMesh::SurfaceMesh()
 {
-    oprops_.push_back();
-
     // allocate standard properties
     // same list is used in operator=() and assign()
     vpoint_ = add_vertex_property<Point>("v:point");
@@ -31,7 +29,6 @@ SurfaceMesh& SurfaceMesh::operator=(const SurfaceMesh& rhs)
     if (this != &rhs)
     {
         // deep copy of property containers
-        oprops_ = rhs.oprops_;
         vprops_ = rhs.vprops_;
         hprops_ = rhs.hprops_;
         eprops_ = rhs.eprops_;
@@ -63,8 +60,6 @@ SurfaceMesh& SurfaceMesh::assign(const SurfaceMesh& rhs)
     if (this != &rhs)
     {
         // clear properties
-        oprops_.clear();
-        oprops_.resize(1);
         vprops_.clear();
         hprops_.clear();
         eprops_.clear();
@@ -109,7 +104,6 @@ SurfaceMesh& SurfaceMesh::assign(const SurfaceMesh& rhs)
 void SurfaceMesh::clear()
 {
     // remove all properties
-    oprops_.clear();
     vprops_.clear();
     hprops_.clear();
     eprops_.clear();
@@ -119,7 +113,6 @@ void SurfaceMesh::clear()
     free_memory();
 
     // add the standard properties back
-    oprops_.push_back();
     vpoint_ = add_vertex_property<Point>("v:point");
     vconn_ = add_vertex_property<VertexConnectivity>("v:connectivity");
     hconn_ = add_halfedge_property<HalfedgeConnectivity>("h:connectivity");
@@ -138,7 +131,6 @@ void SurfaceMesh::clear()
 void SurfaceMesh::free_memory()
 {
     vprops_.free_memory();
-    oprops_.free_memory();
     hprops_.free_memory();
     eprops_.free_memory();
     fprops_.free_memory();
@@ -146,7 +138,6 @@ void SurfaceMesh::free_memory()
 
 void SurfaceMesh::reserve(size_t nvertices, size_t nedges, size_t nfaces)
 {
-    oprops_.reserve(1);
     vprops_.reserve(nvertices);
     hprops_.reserve(2 * nedges);
     eprops_.reserve(nedges);